+++ src/features/reasoning/assets/DeepseekIcon.tsx
